+++ src/_components/index.ts
@@ -6,11 +6,13 @@ export { default as Checkbox } from './checkbox/Checkbox';
 export { default as DiagonalSplitView } from './diagonalSplitView/DiagonalSplitView';
 export { default as ExternalLink } from './externalLink/ExternalLink';
 export { default as FocusAwareStatusBar } from './statusBar/FocusAwareStatusBar';
+export { default as HtmlRenderer } from './htmlRenderer/HtmlRenderer';
 export { default as Icon } from './icon/Icon';
 export { default as LinkList } from './linkList/LinkList';
 export { default as NavigationBar } from './navigation/NavigationBar';
 export { default as NavigationButton } from './navigation/NavigationButton';
 export { default as PillButton } from './pillButton/PillButton';
+export { default as Points } from './points/Points';
 export { default as Reward } from './reward/Reward';
 export { default as RewardImage } from './reward/RewardImage';
 export { default as RewardLoader } from './reward/Reward.loading';
