+++ src/shopDetail/style.ts
@@ -20,11 +20,3 @@ export const PointContainer = styled.View`
 export const Organizer = styled(Typography)`
   margin-bottom: 24px;
 `;
-
-export const Section = styled.View`
-  margin-bottom: ${theme.common.defaultSpacing}px;
-`;
-
-export const SectionTitle = styled(Typography)`
-  margin-bottom: 8px;
-`;
